+++ apps/counter-react-ts-context/jest.config.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+  ...require("../../jest.config.base.cjs"),
+  transform: {
+    "^.+\\.tsx?$": ["ts-jest", { tsconfig: "tsconfig.test.json" }],
+  },
+};