+++ apps/counter-dom-js/CHANGELOG.md
@@ -1,5 +1,26 @@
 # counter-dom-js
 
+## 1.3.0
+
+### Patch Changes
+
+- Updated dependencies [80692ad]
+- Updated dependencies
+- Updated dependencies [d820877]
+- Updated dependencies [d820877]
+  - @lauf/store-follow@1.3.0
+  - @lauf/store@1.3.0
+
+## 1.3.0-alpha.1
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [d820877]
+- Updated dependencies [d820877]
+  - @lauf/store-follow@1.3.0-alpha.1
+  - @lauf/store@1.3.0-alpha.1
+
 ## 1.0.1-alpha.1
 
 ### Patch Changes